Change interface search algorithm
Instead of using the uncertain and weird (albeit maybe fast)
contour function of matplotlib the new algorithm scans the system
layer by layer for boundary cells, determined through how many
neighbouring cells (and if itself) pass a cut-off value which
defaults to the midpoint between the max and min values of the
droplet parameter in the system.

For each layer the edges are yielded as a 2-tuple, which means
that just a few layers can be retrieved if desired, instead of
searching through the entire data structure at once. This
should make implementing a spreading radius function trivial.