Move `get_interface` functions to `interface` module
This unifies all the interface functions into a single module.

The naming of `radius` for interface calculation has been renamed
to `cutoff_radius` which better reflects its purpose. It is now
also set to None by default and if so calculated as the minimum
distance between cells in the system.