Add decorator for drawing graphs
The decorate_graph function absorbs figure drawing commands and sets
titles, limits and other figure options for a called plotting command.
It optionally shows the graph.

Matplotlib is only called from this module, meaning that it hopefully
should not interfere with non-X11 environments as long as plotting
commands are not specified.