First attempt at creating a CLI interface for contact line extraction
The performance is terribad for extracting the cells in large files.
Presumably related to running the conditional on a large data set?
Consider alternative ways, like a limited selection of data.

Also contains an ugly hack for sorting the output data according to the
old data format standard, which is row-column order. This should be
moved from strata to droplets module, but I'm in a rush for this commit.

The strata function is ugly and lacks testing.