interface: add area calculation tool
The interface area is calculated from .xvg files by using trapezoids as
an approximation between points. This means that some area from the top
and bottom of the extraction is cut off, but to correct for that a few
additional and boring inputs would be required so screw that. As long as
cells are kept as a reasonably small size it should be fine regardless.