strata angle: assert that left and right interface split is even
This is done in the case of an uneven set of interface coordinates. Now
the right side picks from the bottom up, just as the left side picks
from the top down. This ascertains both that the right side gets the
explicit and correct number of points, equal to the left side, and that
they correspond to the lowest parts of the interface.