droplets resample: correct array slicing in combining bins
The previous implementation forgot that the data from the grid creation
is stored in (row,column) (i.e y-x) order, which was not used in the
loop for combining bins. A test has been implemented to ensure that bin
slices are implemented correctly, by combining bins into a (3,2) system
and accessing the data from it as it should be.