droplets flow: add `lims` method to limit data in FlowData
This method returns only bins whose value for the input label is within
the input limits. The implementation is very simple but mostly
implemented as a method for easy chaining of different lims when cutting
the data. I also think it's a general enough method to be of use as a
method and not as a function in a separate module.

One big issue is that since FlowData has no concept of which input
labels correspond to coordinates, it cannot determine when a cut along
system axes has occured and return data with a definite grid size.
Hence `shape` and `size` will not be set, since there's no guarantuee
that the cut data corresponds to a regular grid. Whether or not this is
a big issue, I don't know right now since my usage of the data type is
very focused on the user having to keep track of things explicitly when
it's important for them.