FlowData: rename and restructure `size` and `bin_size` properties
`bin_size` -> `spacing` which is the more general term for structured
grids.
`size` -> `origin`, with the additional change that only the starting
coordinates in each dimension is saved since the end coordinate can be
easily derived. This is also more in line with general structured grid
descriptions.

Tests have been updated and hopefully everything should be fixed.