strata: add version as global attribute
Many output functions inside `strata` collected the version number on
their own which was inelegant and stupid since the package was renamed
and these functions all broke individually. Now the version number is
stored as an attribute accessed as `strata.strata.version` which is
still not very elegant, but at least more than a bit better.