strata: Fix bug in contact line output if no std is calculated
When sampling values the `sample_value` function returns both the mean value
and std of the calculation. However, the function that samples data for
the contact line edges (`strata contact_line`) tries to write both the
value and std to the file. In the case of sampling the flow angle
(`flow_angle`) the std is None, but tried to be written as if it were a
float which raises an Exception.

This fix makes sure that only the value is output, not the std. The
latter can be added if necessary.